requirements.txt
opencv-python
numpy
pillow
pyttsx3
streamlit
